requirements.txt
langchain-google-genai
langchain-pinecone
langchain-community
langgraph
pinecone-client>=3.2.2,<4.0.0
python-telegram-bot==21.0
python-dotenv==1.0.1
pypdf==4.2.0
tavily-python==0.3.1
pygithub==2.2.0
flask==3.0.3 